クラス継承とファイル分割完了
